+++ protocol/src/packets/game/clientbound/container_set_slot_packet.rs
@@ -2,12 +2,11 @@
 use crate::packet::*;
 use crate::network::*;
 use bytes::BytesMut;
-use uuid::Uuid;
 use crate::result::*;
 
 pub struct ContainerSetSlotPacket {
-    pub containerId: i32,
-    pub slot: i32,
+    pub containerId: u8,
+    pub slot: i16,
     pub itemStack: ItemStack,
 }
